requirements.txt
streamlit
pandas
plotly
yfinance
streamlit-option-menu
psycopg2-binary
bcrypt
sqlalchemy